Add RobotsController
