+++ applications/dashboard/controllers/class.utilitycontroller.php
@@ -8,37 +8,13 @@
  * @since 2.0
  */
 
-use Garden\EventManager;
-use Gdn_Session as SessionInterface;
-use Vanilla\Contracts\ConfigurationInterface;
-use Vanilla\Web\Robots;
-
 /**
  * Handles /utility endpoint.
  */
 class UtilityController extends DashboardController {
 
-    /** Default robots.txt contents. */
-    private const ROBOTS_DEFAULT = <<<ROBOTS_DEFAULT
-User-agent: *
-Disallow: /entry/
-Disallow: /messages/
-Disallow: /profile/comments/
-Disallow: /profile/discussions/
-Disallow: /search/
-ROBOTS_DEFAULT;
-
     /** A flag used to indicate the site was put into maintenance mode in an automated fashion. */
     const MAINTENANCE_AUTO = 2;
-
-    /** @var ConfigurationInterface */
-    private $configuration;
-
-    /** @var EventManager */
-    private $eventManager;
-
-    /** @var SessionInterface */
-    private $session;
 
     /** @var array Models to automatically instantiate. */
     public $Uses = ['Form'];
@@ -62,16 +38,8 @@ ROBOTS_DEFAULT;
 
     /**
      * UtilityController constructor.
-     *
-     * @param ConfigurationInterface $configuration
-     * @param EventManager $eventManager
-     * @param SessionInterface $session
      */
-    public function __construct(ConfigurationInterface $configuration, EventManager $eventManager, SessionInterface $session) {
-        $this->configuration = $configuration;
-        $this->eventManager = $eventManager;
-        $this->session = $session;
-
+    public function __construct() {
         parent::__construct();
         $this->setHeader('Cache-Control', \Vanilla\Web\CacheControlMiddleware::NO_CACHE);
     }
@@ -668,47 +636,6 @@ ROBOTS_DEFAULT;
         $this->deliveryType(DELIVERY_TYPE_DATA);
         $this->deliveryMethod(DELIVERY_METHOD_JSON);
         $this->render();
-    }
-
-    /**
-     * Get initial rules for the body of robots.txt.
-     *
-     * @return string
-     */
-    private function robotRules(): string {
-        // Config lookup is backwards-compatible with Sitemaps addon.
-        $rules = $this->configuration->get("Robots.Rules", $this->configuration->get("Sitemap.Robots.Rules", null));
-        if ($rules === null) {
-            $rules = self::ROBOTS_DEFAULT;
-        }
-        return $rules;
-    }
-
-    /**
-     * Generate the robots.txt body.
-     */
-    public function robots() {
-        // Clear the session to mimic a crawler.
-        $this->session->UserID = 0;
-        $this->session->User = false;
-
-        $this->deliveryMethod(DELIVERY_METHOD_TEXT);
-        $this->deliveryType(DELIVERY_TYPE_VIEW);
-
-        $isInvisible = $this->configuration->get("Robots.Invisible");
-
-        $robots = new Robots();
-        if ($isInvisible) {
-            $robots->addRule(Robots::ROBOTS_INVISIBLE);
-        } else {
-            $robots->addRule($this->robotRules());
-            $this->eventManager->fire("robots_init", $robots);
-        }
-
-        $this->setHeader("Content-Type", "text/plain");
-        $this->setData("rules", $robots->getRules());
-        $this->setData("sitemaps", $robots->getSitemaps());
-        $this->render("robots");
     }
 
     /**
