+++ applications/dashboard/settings/structure.php
@@ -1000,7 +1000,8 @@ if ($mobileInputFormatter === "Rich" && $richEditorEnabled === false) {
 }
 
 Gdn::router()->setRoute('apple-touch-icon.png', 'utility/showtouchicon', 'Internal');
-Gdn::router()->setRoute("robots.txt", "/utility/robots", "Internal");
+Gdn::router()->setRoute("robots.txt", "/robots", "Internal");
+Gdn::router()->setRoute("utility/robots", "/robots", "Internal");
 
 // Migrate rules from Sitemaps addon.
 if (Gdn::config()->get("Robots.Rules") === false && $sitemapsRobotsRules = Gdn::config()->get("Sitemap.Robots.Rules")) {
